requirements.txt
asgiref==3.11.1
channels==4.3.2
channels_redis==4.3.0
dj-database-url==3.1.2
Django==5.1.4
django-cors-headers==4.6.0
django-extensions==3.2.3
django-sslserver==0.22
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
gunicorn==25.1.0
msgpack==1.1.2
packaging==26.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
redis==7.2.0
sqlparse==0.5.5
whitenoise==6.12.0
